poco: the Zip module depends on XML, Net and Util modules
Without this, in a configuration where the Zip module is selected but
not the XML module, the build fails due to missing expat.h. This is
because POCO builds the XML module as soon as the Zip module is
enabled.

This fixes the build breakage visible at:

  http://free-electrons.com/~thomas/buildroot/test-output-2011-12-02/test-920-output.bz2

with the configuration visible at:

  http://free-electrons.com/~thomas/buildroot/test-output-2011-12-02/test-920-config

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>