libecore: fix build of host-libecore when X11 not available
The host-libecore build tries to build the X11 backend. This works if
you have X11 headers/libraries installed on your build machine, but
fails if you don't, and Buildroot shouldn't depend on such things
being installed.

Therefore, we force host-libecore to not build any of the graphical
backends (X, XCB or DirectFB).

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>