mplayer: not supported on sh4a/sh4aeb
The mplayer configure script says that sh4a is not supported,
according to:

 http://autobuild.buildroot.org/results/c6457c2dea7c1df4d8e0f23995ebfb3285a85aed/build-end.log

Apparently, sh4 is supported, but not sh4a, even though it seems that
sh4a is a super-set of sh4. So it would probably not be too difficult
to get mplayer to build/work for sh4a, but we don't have hardware to
test this easily, and we can leave this work to the first developer
who will need mplayer on sh4a: let's only support what the upstream
package supports.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>