nfs-utils: reformat patch set as git patches
As we will need to add more patches to nfs-utils, we need a correct
ordering when applying patches. Therefore, reformat the patches to use
a git format and git naming.

The nfs-utils-dont-mix-flags.patch is no longer needed as it was
patching Makefile.in files that were being regenerated due to the
package having _AUTORECONF = YES. The Makefile.in are properly
regenerated thanks to the nfs-utils-0002-Patch-taken-from-Gentoo.patch
patch.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>