python: use default configure commands to get shared build
Commit 3c90f754961c6d99059c0b6d0a66bc797a3c017b made Python use a
special ./configure command in order to avoid --enable-shared
--disable-static being passed, because it was causing issues when
building certain modules for a 64 bits system.

However, not having a shared libpython2.7 library for the host
prevents the libxml2 Python binding to get built.

So instead, we use the default configure command, but we add
--enable-static which is needed for Python to build correctly.

Note that we tested the build of Python on a 64 bits host as well as
the build of Python for a 64 bits target, and both went fine, with all
modules built properly.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>