Fix building initramfs kernel with appended dtb
The target for building the kernel with a cpio initramfs is not
calling the append dtb commands, creating a final kernel image
without an appended dtb. Instead it needs to call kernel make on
the intermediate target, then call the append dtb commands
to produce the final image as the primary kernel build target
does.

Signed-off-by: Nathan Ford <nford@westpond.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>