package/opkg: bumping version to 0.2.0
opkg has released a new version v0.2.0. The opkg community is also on
the way to switching completely from subversion to git.
The current git repository is placed on Bitbucket and will probably move
once again if there will be find a new home for the project.
https://bitbucket.org/opkg/opkg

The new version 0.2.0 is almost a bugfixing release with a lot of small
changes.
This patch is adding the license info too.

Signed-off-by: Carsten Schoenert <c.schoenert@gmail.com>